requirements.txt
alembic>=1.12.1
blinker>=1.7.0
click>=8.1.7
dnspython>=2.4.2
email-validator>=2.1.0.post1
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-Mail>=0.9.1
Flask-Migrate>=4.0.5
Flask-SQLAlchemy>=3.1.1
Flask-WTF>=1.2.1
Flask-JWT-Extended>=4.5.3
greenlet>=3.0.1
idna>=3.4
itsdangerous>=2.1.2
Jinja2>=3.1.2
Mako>=1.2.4
MarkupSafe>=2.1.3
python-decouple>=3.8
SQLAlchemy>=2.0.22
typing_extensions>=4.8.0
Werkzeug>=2.3.7
WTForms>=3.1.1
pytest>=7.4.3
